requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
mysql-connector-python==8.1.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
sqlalchemy==2.0.23
aiomysql==0.2.0
passlib==1.7.4
bcrypt==4.1.1